requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt 
mysqlclient
celery
django_celery_beat